Extract components from page.tsx
